Refactored heading to change in heading component and added navlink to replace button
